basic http tests, simplify param methods
basic HTTP behavior tests (not entirely passing yet, since it expresses
desired behavior rather than current).

instead of accepting either a ip query parameter or the IP address as
part of the URL, let's simplify and just standardize on a query param.
it's likely slightly less micro-efficient to parse, but I believe the
usability is greater since it's easier to read in examples.