remove caching entirely
While the cache makes a huge difference in handler microbenchmarks (~10x), and a proximate one in real world HTTP benchmarking (10-20% impact on max reqs/second), it's not a big enough difference to trade off the complexity of an additional external dependency.

Additionally, since this is URL based and the content does not change, it should be trivially cachable with an off-the-shelf proxy-cache, so let's focus on that instead as an infrastructure best practice for separation of concerns.